@@ -1,5 +1,6 @@
 {
   "name": "boclips-player",
+  "type": "module",
   "version": "0.1.0",
   "description": "A framework agnostic player for serving Boclips videos",
   "license": "BSD-3-Clause",
@@ -44,28 +45,29 @@
     "sass": "^1.79.3",
     "sass-loader": "^16.0.2",
     "svg-inline-loader": "^0.8.2",
-    "url-template": "^2.0.8",
+    "url-template": "^3.1.1",
     "uuid": "^10.0.0",
     "webpack-bundle-analyzer": "^4.10.2"
   },
   "devDependencies": {
     "@jest/globals": "^29.7.0",
-    "@webpack-cli/serve": "^1.6.0",
+    "@types/uuid": "^10.0.0",
+    "@webpack-cli/serve": "^2.0.5",
     "audit-ci": "^7.1.0",
     "axios-logger": "^2.8.1",
     "axios-mock-adapter": "^2.0.0",
     "boclips-js-security": "^13.0.1",
     "clean-webpack-plugin": "^4.0.0",
-    "copy-webpack-plugin": "^11.0.0",
+    "copy-webpack-plugin": "^12.0.2",
     "css-loader": "^7.1.2",
     "eventemitter3": "^5.0.1",
     "file-loader": "^6.2.0",
-    "html-webpack-plugin": "^4.5.2",
+    "html-webpack-plugin": "^5.6.0",
     "jest": "^29.7.0",
     "mini-css-extract-plugin": "^2.9.1",
     "prettier": "^3.3.3",
     "prettier-package-json": "^2.8.0",
-    "source-map-loader": "^3.0.0",
+    "source-map-loader": "^5.0.0",
     "ts-jest": "^29.2.5",
     "ts-loader": "^9.5.1",
     "tslint": "^6.1.3",
@@ -73,9 +75,9 @@
     "tslint-plugin-prettier": "^2.3.0",
     "typescript": "^5.6.2",
     "typescript-tslint-plugin": "^1.0.2",
-    "webpack": "^5.94.0",
-    "webpack-cli": "^4.9.0",
-    "webpack-dev-server": "^4.7.4",
-    "webpack-merge": "^5.10.0"
+    "webpack": "^5.95.0",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^5.1.0",
+    "webpack-merge": "^6.0.1"
   }
 }
