@@ -1,20 +1,25 @@
-const { merge } = require('webpack-merge');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
+import { merge } from 'webpack-merge';
+import HtmlWebpackPlugin from 'html-webpack-plugin';
 
-const common = require('../../webpack-config/webpack.dev.js');
+import common from '../../webpack-config/webpack.dev.js';
 
-const path = require('path');
+import { resolve } from 'path';
 
-module.exports = merge(common, {
+import { fileURLToPath } from 'url';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+
+export default merge(common.default, {
   plugins: [
     new HtmlWebpackPlugin({
-      template: path.resolve(__dirname, 'index.html'),
+      template: resolve(__dirname, 'index.html'),
       inject: 'head',
     }),
     new HtmlWebpackPlugin({
       inject: false,
       filename: 'silent-check-sso.html',
-      template: path.resolve(__dirname, 'silent-check-sso.html'),
+      template: resolve(__dirname, 'silent-check-sso.html'),
     }),
   ],
 });
