@@ -1,7 +1,8 @@
 // For a detailed explanation regarding each configuration property, visit:
 // https://jestjs.io/docs/en/configuration.html
+import { v4 as uuidv4 } from 'uuid';
 
-module.exports = {
+export default {
   // All imported modules in your tests should be mocked automatically
   // automock: false,
 
@@ -78,7 +79,7 @@ module.exports = {
     '\\.(css|scss|sass)$': '<rootDir>/__mocks__/styleMock.ts',
     '\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$':
       '<rootDir>/__mocks__/fileMock.ts',
-    uuid: require.resolve('uuid'),
+    uuid: uuidv4,
   },
 
   // An array of regexp pattern strings, matched against all module paths before considered 'visible' to the module loader
