@@ -1,10 +1,8 @@
-const BundleAnalyzerPlugin = require('webpack-bundle-analyzer')
-  .BundleAnalyzerPlugin;
+import { BundleAnalyzerPlugin } from 'webpack-bundle-analyzer';
+import { merge } from 'webpack-merge';
+import production from './webpack.prod.js';
 
-const { merge } = require('webpack-merge');
-const production = require('./webpack.prod.js');
-
-module.exports = merge(production, {
+export default merge(production, {
   plugins: [
     new BundleAnalyzerPlugin({
       analyzerMode: 'static',
