@@ -1,11 +1,16 @@
-const { CleanWebpackPlugin } = require('clean-webpack-plugin');
-const path = require('path');
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
+import { CleanWebpackPlugin } from 'clean-webpack-plugin';
+import path from 'path';
+import MiniCssExtractPlugin from 'mini-css-extract-plugin';
+
+import { fileURLToPath } from 'url';
+
+const __filename = fileURLToPath(import.meta.url); // get the resolved path to the file
+const __dirname = path.dirname(__filename); // get the name of the directory
 
 const srcPath = path.resolve(__dirname, '../src');
 const distPath = path.resolve(__dirname, '../dist');
 
-module.exports = {
+export default {
   entry: path.resolve(srcPath, 'index.ts'),
   output: {
     filename: 'index.js',
@@ -16,7 +21,6 @@ module.exports = {
     libraryTarget: 'umd',
   },
   resolve: {
-    fallback: { querystring: require.resolve('querystring-es3') },
     alias: {
       src: srcPath,
     },
